Multimodal-QA-Agent/test_agent.py
import unittest
import os
import base64
from PIL import Image
from io import BytesIO
from dotenv import load_dotenv
from multimodal_agent import MultimodalQAAgent

class TestMultimodalAgent(unittest.TestCase):
    """Test cases for the Multimodal QA Agent."""
    
    @classmethod
    def setUpClass(cls):
        """Set up test environment."""
        load_dotenv()
        cls.agent = MultimodalQAAgent()
        
        # Create a simple test image
        cls.test_image = cls.create_test_image()
    
    @staticmethod
    def create_test_image():
        """Create a simple test image."""
        img = Image.new('RGB', (200, 200), color='red')
        buffer = BytesIO()
        img.save(buffer, format='JPEG')
        return base64.b64encode(buffer.getvalue()).decode()
    
    def test_agent_initialization(self):
        """Test that the agent initializes correctly."""
        self.assertIsNotNone(self.agent.vision_model)
        self.assertIsNotNone(self.agent.text_model)
        self.assertIsNotNone(self.agent.api_key)
    
    def test_text_only_question(self):
        """Test text-only question answering."""
        question = "What should I look for in a landscape photo?"
        response = self.agent.ask_question_text_only(question)
        
        self.assertIsInstance(response, str)
        self.assertGreater(len(response), 0)
        print(f"Text-only response: {response[:100]}...")
    
    def test_vision_question(self):
        """Test vision-based question answering."""
        question = "What color is dominant in this image?"
        
        try:
            response = self.agent.ask_question(question, self.test_image)
            self.assertIsInstance(response, str)
            self.assertGreater(len(response), 0)
            print(f"Vision response: {response[:100]}...")
        except Exception as e:
            print(f"Vision test failed (expected with test image): {e}")
            # This might fail with our simple test image, which is expected
    
    def test_model_info(self):
        """Test getting model information."""
        info = self.agent.get_model_info()
        
        self.assertIn("vision_model", info)
        self.assertIn("text_model", info)
        self.assertEqual(info["provider"], "Google Generative AI")
        self.assertEqual(info["framework"], "LangChain")
    
    def test_comprehensive_analysis(self):
        """Test comprehensive image analysis."""
        try:
            results = self.agent.analyze_image_comprehensive(self.test_image)
            self.assertIsInstance(results, dict)
            self.assertGreater(len(results), 0)
            print(f"Comprehensive analysis completed with {len(results)} results")
        except Exception as e:
            print(f"Comprehensive analysis failed (expected with test image): {e}")

if __name__ == '__main__':
    # Run the tests
    unittest.main(verbosity=2)
