Multimodal-QA-Agent/requirements.txt
streamlit==1.28.2
langchain==0.1.4
langchain-google-genai==0.0.8
google-generativeai==0.3.2
pillow==10.1.0
requests==2.31.0
python-dotenv==1.0.0
